feat: add a common TopicClient with publish and subscribe methods
Add a common TopicClient.

Remove the old implementation-specific topic clients.

Add a publish method that takes a ByteArray value.

Add a subscribe method that returns a flow of TopicMessages. I'm not
convinced that the common/android/jvm split is correct here since the
flow creation logic is duplicated. I'll experiment with containing the
grpc code as much as possible, so it can be shared.

Add tests for publish and subscribe.